File System and END
ADding Multer for the file system
changing all the delete, put and add object
END of this project

+++ P6 - API/ProjectP6/backend/controllers/stuff.js
@@ -1,28 +1,63 @@
 const Thing = require('../models/Thing');
+const fs = require('fs'); //File System
 
 //Create an Object
-exports.creatThing = (req, res, next) => {
-    delete req.body._id; //Delete the Id from the Thing
-    const thing = new Thing({ //Creating a New Thing
-        ...req.body //Get all the information from the Thing Model
+exports.createThing = (req, res, next) => {
+    const thingObject = JSON.parse(req.body.thing);
+    delete thingObject._id;
+    delete thingObject._userId;
+    const thing = new Thing({
+        ...thingObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
+
     thing.save()
-        .then(() => res.status(201).json({ message: 'Objet enregistre !' }))
-        .catch(error => res.status(400).json({ error }));
+        .then(() => { res.status(201).json({ message: 'Objet enregistré !' }) })
+        .catch(error => { res.status(400).json({ error }) })
 };
 
 //Modify an Object
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-        .catch(error => res.status(400).json({ error }));
+    const thingObject = req.file ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete thingObject._userId;
+    Thing.findOne({ _id: req.params.id })
+        .then((thing) => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié!' }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
 
 //Delete an Object
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+        .then(thing => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non-autorise' });
+            } else {
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Objet supprime !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
 
 //Get One Object
